pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.12.39",
    "libsql>=0.1.11",
    "libsql-client>=0.3.1",
    "psycopg2-binary>=2.9.11",
    "fastsql>=2.1.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
